fix errors reported by PHPStan

+++ classes/Imagifybeat/Actions.php
@@ -301,7 +301,7 @@ final class Actions {
 			// Sanitize the IDs: IDs come as strings, prefixed with an undescore character (to prevent JavaScript from screwing everything).
 			$media_ids = array_keys( $media_statuses );
 			$media_ids = array_map( function( $media_id ) {
-				return (int) substr( $media_id, 1 );
+				return substr( $media_id, 1 );
 			}, $media_ids );
 			$media_ids = array_filter( $media_ids );
